Obtains the labels for a list of terms

Attempts to obtain the label for each term, identified by IRI, in the input list. Terms for which no label is found in the database will have NA as the label in the result (see Value).

get_term_label(term_iris, preserveOrder = FALSE, verbose = FALSE)

Arguments

term_iris
character, a list of term IRIs

preserveOrder
logical, whether the resulting data frame (see Value) is to be in the same row order as termIRIs. The default is not to preserve order.

verbose
logical, whether to print information about possibly time-consuming operations.

Value

A data.frame with columns "id" and "label". The "id" column contains the IRIs. The label will be NA for term IRIs that are not present in the KB, or for which the KB cannot produce a label.