Determine whether two phenotypes are mutually exclusive

Determines whether the data in the KB includes evidence for mutual exclusivity between two given phenotypes.

mutual_exclusivity_pairwise(
  phenotype.a,
  phenotype.b,
  studies = NULL,
  charstates = NULL
)

Arguments

phenotype.a

character or phenotype object, phenotype ID (IRI) if character, or a phenotype object obtained by passing phenotype ID to as.phenotype() function.

phenotype.b

character or phenotype object, phenotype ID (IRI) if character, or a phenotype object obtained by passing phenotype ID to as.phenotype() function.

studies

character, a vector of study IDs. This is an optional parameter that acts as a filter in case the determination of mutual exclusivity is to be based only on the evidence found in a particular set of studies. The default is NULL, which means that all studies present in the KB database will be considered for evidence.

charstates

dataframe, a dataframe obtained from charstates() by passing a list of the two phenotypes as argument. This optional parameter exists to speed up the computation for the mutually_exclusive() function since mutually_exclusive() repeatedly calls mutual_exclusivity_pairwise(). The default is NULL. Note that passing this argument but doing so incorrectly can result in wrong output.

Value

A character (string), the mutual exclusivity type among the two phenotypes. See mutually_exclusive() for documentation on the possible values, although note that this function returns these as a character vector, not levels of an ordered factor.

Details

See mutually_exclusive() for details on which links in the data are considered as strong or weak evidence for mutual exclusivity and compatibility.