Query the list of studies by taxa, anatomical entities, and qualities.

Return studies that contain taxa which are members of the optional input taxon, and characters which have phenotype annotations subsumed by the given entity and quality terms.

pk_get_study_list(taxon = NA, entity = NA, quality = NA,
  phenotype = NA, includeRels = NA, relation = "part of")

Arguments

taxon
character. The name of the taxon by which to filter, if any.

entity
character. The name of the anatomical entity by which to filter, if any.

quality
character. The name of the phenotypic quality by which to filter, if any.

phenotype
character. The phenotype (as its identifier) by which to filter, if any. If provided, matching studies (through its one or more of its character states) must be linked to the given phenotype, or one subsumed by it. This must be provided as identifier, no text search will be performed for resolution.

includeRels
character or vector of characters. The names of relationships for anatomical entities to include in addition to subtype (is_a, rdfs:subClassOf). Defaults to "part_of". Set to FALSE to not include any additional relationships. Otherwise one or more of "part of", "historical homologous to", and "serially homologous to", or set to TRUE to include all possible ones. It is acceptable to use unambiguous prefixes, for example "historical homolog".

relation
character. Deprecated, for backwards compatibility defaults to part of. Only used if includeRels is left at its default value.

Value

data.frame

Examples

if (FALSE) {
# by default, parts are included
slist <- pk_get_study_list(taxon = "Siluridae", entity = "fin")
colnames(slist)
nrow(slist)

# can also disable parts
slist <- pk_get_study_list(taxon = "Siluridae", entity = "fin", includeRels = FALSE)
nrow(slist)

# or filter studies only by entity, including their parts
slist <- pk_get_study_list(entity = "pelvic fin", includeRels = c("part of"))
nrow(slist)

# or filter studies only by entity, including their parts
slist <- pk_get_study_list(entity = "pelvic fin", includeRels = c("part of"))
nrow(slist)

# including not only parts but also historical and serial homologs
slist <- pk_get_study_list(entity = "pelvic fin",
                           includeRels = c("part of",
                                           "serially homologous to",
                                           "historical homologous to"))
nrow(slist)
# relationship names can be given as prefixes
slist1 <- pk_get_study_list(entity = "pelvic fin",
                            includeRels = c("part", "serial", "historical"))
nrow(slist1) == nrow(slist)

# or apply no filter, obtaining all studies in the KB
slist <- pk_get_study_list()
nrow(slist)
}