Get term details (ID, label, definition)

Retrieve details about a taxon, an anatomical structure, a gene, or a phenotypic quality.

pk_taxon_detail(term, verbose = FALSE)

pk_anatomical_detail(term, verbose = FALSE)

pk_phenotype_detail(term, verbose = FALSE)

pk_gene_detail(term, taxon = NA, verbose = FALSE)

Arguments

term
character, the query term, either as name or IRI. Names are looked up against taxonomies, anatomy ontologies, and PATO for pk_taxon_detail, pk_anatomical_detail, and pk_phenotype_detail, respectively.

For pk_taxon_detail this can also be a list (or vector) of terms (taxa).

verbose
logical, whether informative messages should be printed. The default is FALSE.

taxon
character, the NCBI taxon name or corresponding NCBITaxon ontology IRI for which to match the gene name.

Value

A data.frame, with at least columns "id" and "label".

For pk_taxon_detail, additional columns are "extinct" (logical), "rank.id", "rank.label", and where available "common_name". The rows corresponding to taxon names that failed to be resolved to IRIs will be NA.

For pk_anatomical_detail and pk_phenotype_detail, the additional column is "definition".

For pk_gene_detail, the additional columns are "taxon.id" and "taxon.label" for the corresponding NCBI Taxonomy ID and name, and "matchType" ('exact' or 'partial').

Examples

pk_taxon_detail("Coralliozetus")

#>                                           id         label extinct
#> 1 http://purl.obolibrary.org/obo/VTO_0042955 Coralliozetus   FALSE
#>                                          rank.id rank.label common_name
#> 1 http://purl.obolibrary.org/obo/TAXRANK_0000005      genus        <NA>

pk_anatomical_detail("basihyal bone")

#>           label                               isDefinedBy
#> 1 basihyal bone http://purl.obolibrary.org/obo/uberon.owl
#>                                                                                 definition
#> 1 Replacement bone that is median and is the anterior-most bone of the ventral hyoid arch.
#>                                              id
#> 1 http://purl.obolibrary.org/obo/UBERON_0011618

pk_gene_detail("socs5")

#>                                                  id  label matchType
#> 1             http://xenbase.org/XB-GENEPAGE-479592  socs5     exact
#> 2 http://www.informatics.jax.org/marker/MGI:2385459  Socs5     exact
#> 3               http://zfin.org/ZDB-GENE-061013-408 socs5a   partial
#> 4                http://zfin.org/ZDB-GENE-080722-18 socs5b   partial
#>                                         taxon.id  taxon.label
#> 1  http://purl.obolibrary.org/obo/NCBITaxon_8353      Xenopus
#> 2 http://purl.obolibrary.org/obo/NCBITaxon_10090 Mus musculus
#> 3  http://purl.obolibrary.org/obo/NCBITaxon_7955  Danio rerio
#> 4  http://purl.obolibrary.org/obo/NCBITaxon_7955  Danio rerio