Deprecated functions in rphenoscape.

The functions listed below are deprecated and will become defunct in the near future. When possible, alternative functions with identical or similar functionality are mentioned. Where the signature of the alternative (new) function differs significantly from that of the deprecated one, help pages for the deprecated functions are available at help("<function>-deprecated").

pk_get_iri(...)

pk_get_study_list(...)

pk_taxon_class(x, verbose = TRUE)

pk_anatomical_class(x, verbose = TRUE)

pk_phenotype_class(x, verbose = TRUE)

pk_get_ontotrace_xml(
  taxon,
  entity,
  relation = "part of",
  variable_only = TRUE,
  strict = TRUE
)

pk_get_study_xml(study_ids)

pk_is_descendant(term, candidates, includeRels = c("none", "part_of"))

pk_is_ancestor(term, candidates, includeRels = c("none", "part_of"))

pk_taxon_detail(term, verbose = FALSE)

pk_anatomical_detail(term, verbose = FALSE)

pk_phenotype_detail(term, verbose = FALSE)

pk_gene_detail(term, taxon = NA, verbose = FALSE)

pk_is_extinct(taxon, verbose = FALSE)

pk_get_ontotrace(nex)

pk_get_ontotrace_meta(nex)

pk_get_study(nexmls)

pk_get_study_meta(nexmls)

pk_get_iri

For pk_get_iri() use get_term_iri() instead.

pk_get_study_list

For pk_get_study_list() use get_studies() instead. Note that get_studies() returns an empty result set if nothing was found, in contrast to pk_get_study_list, which returns FALSE in that case (and prints a message).

pk_taxon_class

For pk_taxon_class() use term_classification() with as="taxon" instead. Note that pk_taxon_class() defaults verbose to TRUE instead of FALSE as term_classification() does. Also pk_taxon_class() return value has the legacy @id name.

pk_anatomical_class

For pk_anatomical_class() use term_classification() with as="anatomy" instead. Note that pk_anatomical_class() defaults verbose to TRUE instead of FALSE as term_classification() does. Also pk_anatomical_class() return value has the legacy @id name.

pk_phenotype_class

For pk_phenotype_class() use term_classification() with as="pato" instead. Note that pk_phenotype_class() defaults verbose to TRUE instead of FALSE as term_classification() does. Also pk_phenotype_class() return value has the legacy @id name.

pk_get_ontotrace_xml

For pk_get_ontotrace_xml() use get_ontotrace_data() instead.

pk_get_study_xml

For pk_get_study_xml() use get_study_data() instead. Note that pk_get_study_xml() maintains legacy functionality of displaying messages. get_study_data() has a verbose argument to control displaying messages.

pk_is_descendant

For pk_is_descendant() use is_descendant() instead.

pk_is_ancestor

For pk_is_ancestor() use is_ancestor() instead.

pk_taxon_detail

For pk_taxon_detail() use taxon_info() instead.

pk_anatomical_detail

For pk_anatomical_detail() use anatomy_term_info() instead.

pk_phenotype_detail

For pk_phenotype_detail() use phenotypic_quality_term_info() instead.

pk_gene_detail

For pk_gene_detail() use gene_info() instead.

pk_is_extinct

For pk_is_extinct() use is_extinct() instead.

pk_get_ontotrace

For pk_get_ontotrace() use get_char_matrix() instead.

pk_get_ontotrace_meta

For pk_get_ontotrace_meta() use get_char_matrix_meta() instead.

pk_get_study

For pk_get_study() use get_char_matrix() with lapply instead. Note that pk_get_study() always displays messages instead of using the verbose parameter as get_char_matrix() does.

pk_get_study_meta

For pk_get_study_meta() use get_char_matrix_meta() with lapply instead.