Get term details (ID, label, definition)

Retrieve details about a taxon, an anatomical structure, a gene, or a phenotypic quality.

taxon_info(term, verbose = FALSE)

anatomy_term_info(term, verbose = FALSE)

phenotypic_quality_term_info(term, verbose = FALSE)

gene_info(term, taxon = NA, verbose = FALSE)

Arguments

term

character, the query term, either as name or IRI. Names are looked up against taxonomies, anatomy ontologies, and PATO for taxon_info, anatomy_term_info, and phenotypic_quality_term_info, respectively.

For taxon_info this can also be a list (or vector) of terms (taxa).

verbose

logical, whether informative messages should be printed. The default is FALSE.

taxon

character, the NCBI taxon name or corresponding NCBITaxon ontology IRI for which to match the gene name.

Value

A data.frame, with at least columns "id" and "label".

For taxon_info, additional columns are "extinct" (logical), "rank.id", "rank.label", and where available "common_name". The rows corresponding to taxon names that failed to be resolved to IRIs will be NA.

For anatomy_term_info and phenotypic_quality_term_info, the additional column is "definition".

For gene_info, the additional columns are "taxon.id" and "taxon.label" for the corresponding NCBI Taxonomy ID and name, and "matchType" ('exact' or 'partial').

Examples

taxon_info("Coralliozetus")
#>                                           id         label extinct
#> 1 http://purl.obolibrary.org/obo/VTO_0042955 Coralliozetus   FALSE
#>                                          rank.id rank.label common_name
#> 1 http://purl.obolibrary.org/obo/TAXRANK_0000005      genus        <NA>
anatomy_term_info("basihyal bone")
#>           label                               isDefinedBy
#> 1 basihyal bone http://purl.obolibrary.org/obo/uberon.owl
#>                                                                                 definition
#> 1 Replacement bone that is median and is the anterior-most bone of the ventral hyoid arch.
#>                                              id
#> 1 http://purl.obolibrary.org/obo/UBERON_0011618
gene_info("socs5")
#>                                                     id  label matchType
#> 1 http://www.informatics.jax.org/accession/MGI:2385459  Socs5     exact
#> 2                  http://zfin.org/ZDB-GENE-061013-408 socs5a   partial
#> 3                   http://zfin.org/ZDB-GENE-080722-18 socs5b   partial
#>                                         taxon.id  taxon.label
#> 1 http://purl.obolibrary.org/obo/NCBITaxon_10090 Mus musculus
#> 2  http://purl.obolibrary.org/obo/NCBITaxon_7955  Danio rerio
#> 3  http://purl.obolibrary.org/obo/NCBITaxon_7955  Danio rerio